When resolving $refs, retain existing properties.
It's quite common when using $ref to re-use a definition that you'll
want to provide a description about the context in which a type is
used. For example:

{"definitions": {
    "timestamp": {
        "type": "integer",
        "description": "Some unix timestamp"},
    "account": {
        "registrationTime": {
            "$ref": "#/definitions/timestamp",
            "description": "When signup was completed."},
        "emailVerificationTime": {
            "$ref": "#/definitions/timestamp",
            "description": "When email was verified."}}}

Prior to this code change, the descriptions on registrationTime and
emailVerificationTime would have been discarded. With this code change,
we keep the more informative description! Huzzah.